fix: renamed components + small fixes
