+++ src/renderer/entities/transaction/ui/Fee/Fee.tsx
@@ -1,61 +1,20 @@
-import { type ApiPromise } from '@polkadot/api';
-import { BN } from '@polkadot/util';
 import { useUnit } from 'effector-react';
-import { memo, useEffect, useState } from 'react';
+import { memo } from 'react';
 
-import { type Asset, type Transaction } from '@/shared/core';
+import { type Asset } from '@/shared/core';
 import { AssetBalance } from '@/shared/ui-entities';
 import { AssetFiatBalance, priceProviderModel } from '@/entities/price';
-import { transactionService } from '../../lib';
 import { FeeLoader } from '../FeeLoader/FeeLoader';
 
 type Props = {
-  api: ApiPromise | null;
-  multiply?: number;
+  fee: string;
+  isLoading?: boolean;
   asset: Asset;
-  transaction?: Transaction | null;
   className?: string;
-  onFeeChange?: (fee: string) => void;
-  onFeeLoading?: (loading: boolean) => void;
 };
 
-export const Fee = memo(({ api, multiply = 1, asset, transaction, className, onFeeChange, onFeeLoading }: Props) => {
+export const Fee = memo(({ fee, isLoading, asset, className }: Props) => {
   const fiatFlag = useUnit(priceProviderModel.$fiatFlag);
-
-  const [fee, setFee] = useState('');
-  const [isLoading, setIsLoading] = useState(true);
-
-  const updateFee = (fee: string) => {
-    const totalFee = new BN(fee).muln(multiply).toString();
-    setFee(totalFee);
-    onFeeChange?.(totalFee);
-  };
-
-  useEffect(() => {
-    onFeeLoading?.(isLoading);
-  }, [isLoading]);
-
-  useEffect(() => {
-    setIsLoading(true);
-    if (!api || !transaction?.accountId) return;
-    let mounted = true;
-
-    api.isReady
-      .then(() => transactionService.getTransactionFee(transaction, api))
-      .then((fee) => {
-        if (mounted) {
-          updateFee(fee);
-          setIsLoading(false);
-        }
-      })
-      .catch((error) => {
-        console.info('Error getting fee - ', error);
-      });
-
-    return () => {
-      mounted = false;
-    };
-  }, [transaction, api]);
 
   if (isLoading) {
     return <FeeLoader fiatFlag={Boolean(fiatFlag)} />;
