+++ src/renderer/entities/transaction/ui/FeeWithLabel/FeeWithLabel.tsx
@@ -5,14 +5,34 @@ import { useI18n } from '@/shared/i18n';
 import { cnTw } from '@/shared/lib/utils';
 import { DetailRow, FootnoteText } from '@/shared/ui';
 import { Fee } from '../Fee/Fee';
-import { FeeWithoutDataLoading } from '../FeeWithoutDataLoading/Fee';
+import { FeeWithDataLoading } from '../FeeWithDataLoading/FeeWithDataLoading';
 
-type Props = ComponentProps<typeof Fee> & {
+type Props = ComponentProps<typeof FeeWithDataLoading> & {
   label?: string;
   wrapperClassName?: string;
 };
 
-export const FeeWithLabel = ({ label, wrapperClassName, ...feeProps }: Props) => {
+/** @deprecated Use FeeWithLabel with complexTxStore instead */
+export const FeeWithLabelWithDataLoading = ({ label, wrapperClassName, ...feeProps }: Props) => {
+  const { t } = useI18n();
+
+  return (
+    <DetailRow
+      label={<FootnoteText className="text-text-tertiary">{label || t('operation.networkFee')}</FootnoteText>}
+      className={cnTw('text-text-primary', wrapperClassName)}
+      testId={TEST_IDS.OPERATIONS.ESTIMATE_FEE}
+    >
+      <FeeWithDataLoading {...feeProps} />
+    </DetailRow>
+  );
+};
+
+type FeeWithLabelProps = ComponentProps<typeof Fee> & {
+  label?: string;
+  wrapperClassName?: string;
+};
+
+export const FeeWithLabel = ({ label, wrapperClassName, ...feeProps }: FeeWithLabelProps) => {
   const { t } = useI18n();
 
   return (
@@ -22,25 +42,6 @@ export const FeeWithLabel = ({ label, wrapperClassName, ...feeProps }: Props) =>
       testId={TEST_IDS.OPERATIONS.ESTIMATE_FEE}
     >
       <Fee {...feeProps} />
-    </DetailRow>
-  );
-};
-
-type PropsWithoutDataLoading = ComponentProps<typeof FeeWithoutDataLoading> & {
-  label?: string;
-  wrapperClassName?: string;
-};
-
-export const FeeWithLabelWithoutDataLoading = ({ label, wrapperClassName, ...feeProps }: PropsWithoutDataLoading) => {
-  const { t } = useI18n();
-
-  return (
-    <DetailRow
-      label={<FootnoteText className="text-text-tertiary">{label || t('operation.networkFee')}</FootnoteText>}
-      className={cnTw('text-text-primary', wrapperClassName)}
-      testId={TEST_IDS.OPERATIONS.ESTIMATE_FEE}
-    >
-      <FeeWithoutDataLoading {...feeProps} />
     </DetailRow>
   );
 };
