--- src/renderer/entities/transaction/ui/FeeWithoutDataLoading/Fee.tsx
@@ -1,29 +1,0 @@
-import { useUnit } from 'effector-react';
-import { memo } from 'react';
-
-import { type Asset } from '@/shared/core';
-import { AssetBalance } from '@/shared/ui-entities';
-import { AssetFiatBalance, priceProviderModel } from '@/entities/price';
-import { FeeLoader } from '../FeeLoader/FeeLoader';
-
-type Props = {
-  fee: string;
-  isLoading?: boolean;
-  asset: Asset;
-  className?: string;
-};
-
-export const FeeWithoutDataLoading = memo(({ fee, isLoading, asset, className }: Props) => {
-  const fiatFlag = useUnit(priceProviderModel.$fiatFlag);
-
-  if (isLoading) {
-    return <FeeLoader fiatFlag={Boolean(fiatFlag)} />;
-  }
-
-  return (
-    <div className="flex flex-col items-end gap-y-0.5">
-      <AssetBalance value={fee} asset={asset} className={className} />
-      <AssetFiatBalance asset={asset} amount={fee} />
-    </div>
-  );
-});
