+++ src/renderer/entities/transaction/ui/index.ts
@@ -1,4 +1,4 @@
-export { Fee } from './Fee/Fee';
+export { FeeWithDataLoading as Fee } from './FeeWithDataLoading/FeeWithDataLoading';
 export { XcmFee } from './XcmFee/XcmFee';
 export { FeeLoader } from './FeeLoader/FeeLoader';
 export { OperationResult } from './OperationResult/OperationResult';
@@ -8,8 +8,7 @@ export { MultisigDeposit } from './MultisigDeposit/MultisigDeposit';
 export { MultisigDepositWithLabel } from './MultisigDepositWithLabel/MultisigDepositWithLabel';
 export { ProxyDeposit } from './ProxyDeposit/ProxyDeposit';
 export { ProxyDepositLabel } from './ProxyDepositLabel/ProxyDepositLabel';
-export { FeeWithLabel } from './FeeWithLabel/FeeWithLabel';
-export { FeeWithLabelWithoutDataLoading } from './FeeWithLabel/FeeWithLabel';
+export { FeeWithLabelWithDataLoading, FeeWithLabel } from './FeeWithLabel/FeeWithLabel';
 export { XcmFeeWithLabel } from './XcmFeeWithLabel/XcmFeeWithLabel';
 export { TransactionTitle } from './TransactionTitle/TransactionTitle';
 export { DeliveryFee } from './DeliveryFee/DeliveryFee';
