+++ src/renderer/widgets/Staking/Restake/ui/RestakeForm.tsx
@@ -8,7 +8,7 @@ import { formatBalance, toAddress, toShortAddress } from '@/shared/lib/utils';
 import { Button, InputHint, MultiSelect } from '@/shared/ui';
 import { AssetBalance } from '@/shared/ui-entities';
 import { SignatorySelector } from '@/entities/operations';
-import { FeeWithLabel, MultisigDepositWithLabel } from '@/entities/transaction';
+import { FeeWithLabelWithDataLoading, MultisigDepositWithLabel } from '@/entities/transaction';
 import { AccountAddress, ProxyWalletAlert, accountUtils, walletUtils } from '@/entities/wallet';
 import { walletSelect } from '@/aggregates/wallet-select';
 import { AmountInput } from '@/features/assets-balances';
@@ -213,7 +213,7 @@ const FeeSection = () => {
         />
       )}
 
-      <FeeWithLabel
+      <FeeWithLabelWithDataLoading
         label={t('staking.networkFee', { count: shards.value.length || 1 })}
         api={api}
         asset={network.chain.assets[0]}
@@ -223,7 +223,7 @@ const FeeSection = () => {
       />
 
       {transactions && transactions.length > 1 && (
-        <FeeWithLabel
+        <FeeWithLabelWithDataLoading
           label={t('staking.networkFeeTotal')}
           api={api}
           asset={network.chain.assets[0]}
