+++ src/renderer/widgets/UnlockModal/ui/UnlockForm.tsx
@@ -9,7 +9,7 @@ import { Button, InputHint, MultiSelect } from '@/shared/ui';
 import { type DropdownResult } from '@/shared/ui/Dropdowns/common/types';
 import { Address, AssetBalance } from '@/shared/ui-entities';
 import { SignatorySelector } from '@/entities/operations';
-import { FeeWithLabel, MultisigDepositWithLabel } from '@/entities/transaction';
+import { FeeWithLabelWithDataLoading, MultisigDepositWithLabel } from '@/entities/transaction';
 import { ProxyWalletAlert, walletUtils } from '@/entities/wallet';
 import { walletSelect } from '@/aggregates/wallet-select';
 import { AmountInput } from '@/features/assets-balances';
@@ -218,7 +218,7 @@ const FeeSection = () => {
         />
       )}
 
-      <FeeWithLabel
+      <FeeWithLabelWithDataLoading
         label={t('operation.networkFee', { count: shards.value.length || 1 })}
         api={api}
         asset={chain.assets[0]}
@@ -228,7 +228,7 @@ const FeeSection = () => {
       />
 
       {transactions && transactions.length > 1 && (
-        <FeeWithLabel
+        <FeeWithLabelWithDataLoading
           label={t('operation.networkFeeTotal')}
           api={api}
           asset={chain.assets[0]}
